Made MockHttpEntity more obvious what it was doing by separating it from a basic non-implementation
